Gate the monthly maximum LHA to the years it existed
Parameters are backdated to 2015 when the tree is loaded, so leaving the
early years out of the YAML did not stop the monthly maximum applying to
them: a 2019 Central London category E case took the April 2020 figure of
GBP 2,579.98 a month, against GBP 1,917.12 then in force. Universal Credit
now falls back to the weekly Housing Benefit rate before April 2020, which
restores the previous behaviour for those years, and the boundary is tested.

The three rate variables also move into files matching their class names, per
the model-structure guidance, with the shared category-maximum helper beside
the other LHA helpers.
