+++ policyengine_uk/parameters/gov/dwp/LHA/freeze.yaml
@@ -14,7 +14,7 @@ values:
       reference:
         - title: The Rent Officers (Housing Benefit and Universal Credit Functions) (Modification) Order 2020 (determinations in 2021)
           href: https://www.legislation.gov.uk/uksi/2020/1519/made
-        - title: The Rent Officers (Housing Benefit and Universal Credit Functions) (Modification) Order 2021 (determinations in 2022)
+        - title: The Rent Officers (Housing Benefit and Universal Credit Functions) (Amendment and Modification) Order 2021 (determinations in 2022)
           href: https://www.legislation.gov.uk/uksi/2021/1380/made
         - title: The Rent Officers (Housing Benefit and Universal Credit Functions) (Modification) Order 2023 (determinations in 2023)
           href: https://www.legislation.gov.uk/uksi/2023/6/made
