+++ policyengine_uk/parameters/gov/dwp/LHA/maximum_monthly/index.yaml
@@ -3,5 +3,5 @@ description: >-
 metadata:
   label: Maximum monthly LHA
   reference:
-    - title: The Rent Officers (Universal Credit Functions) (Amendment) Order 2015 (omitting the maximum for 2016)
+    - title: The Rent Officers (Housing Benefit and Universal Credit Functions) (Local Housing Allowance Amendments) Order 2015 (omitting the maximum for 2016)
       href: https://www.legislation.gov.uk/uksi/2015/1753/made
